Ship a conformance kit, and stop cutting texture ids in half
From a report by someone porting libgui to bgfx.

A backend for an unusual RHI is ported by hand and a hand port is nearly
right: corners a shade off, a shadow that clips, glyphs half a pixel up
at 1.5x. Nothing compared his port's pixels against a reference, so the
reference ships now. The scene gallery moves from libgui_soft's tests
into the library, with the font passed in rather than embedded;
libgui_enable_reference_render renders any frame on the CPU, and
libgui_conformance_* walks a gallery covering every primitive, so a
failure names the scene and the scene names the primitive. The test that
matters is that the reference handed to C is byte-identical to
libgui_soft's own output, because a kit that certifies the wrong thing is
worse than no kit. The golden and parity tests pass unchanged, which is
the evidence the move was transparent.

Texture ids were cut to 32 bits on the way through, so a native handle or
a pointer came back as a different texture with no error anywhere. They
are u64 in core, the backends and the ABI now, and LibguiBatch grew a
field, so LIBGUI_ABI_VERSION is 2.

libgui_viewport could only show a whole texture, which pooled targets,
half-resolution scenes and packed atlases all break; libgui_viewport_uv
takes a sub-rect, and reversing v turns the image over for bottom-left
APIs -- tested before documented. Vertex attributes now report their
name, pinned against the contract, instead of leaving the mapping
hard-coded on the host side.

The rules a renderer has to agree with were true but unwritten, or
written in a shader comment: blending, colour space, texture format,
sampling, orientation, and the order within a frame that decides whether
a viewport lags while resizing. They are in the header, in one block.

Also: an example in that block nested a block comment and broke the
header. no_comment_in_the_header_ends_early catches it in Rust now,
rather than in the one CI job that has a C compiler.

+++ crates/libgui/src/draw.rs
@@ -6,7 +6,10 @@ use std::ops::Range;
 #[derive(Clone, Copy, Debug, PartialEq, Eq, Hash)]
 pub enum TextureId {
     Atlas,
-    User(u32),
+    /// One of the host's own textures. Opaque to libgui: it is whatever the
+    /// host's renderer uses to name a texture, and 64 bits so that a native
+    /// handle or a pointer fits without being cut down.
+    User(u64),
 }
 
 use crate::render_contract::PrimitiveKind;
